+++ pkg/config/console_task.go
@@ -37,8 +37,8 @@ func (obj *ConsoleTask) IsEnabled() bool {
 func (con *ConsoleTask) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	var boolval bool
 	if err := unmarshal(&boolval); err != nil {
-		type t ConsoleTask
-		var defo = t{}
+		type tConsoleTask ConsoleTask
+		var defo = tConsoleTask{}
 		if err := unmarshal(&defo); err != nil {
 			return err
 		}
